@@ -38,50 +38,50 @@
   "dependencies": {
     "@material-ui/core": "^3.9.3",
     "@material-ui/icons": "^3.0.2",
-    "axios": "^0.18.0",
+    "axios": "^0.18.1",
     "bootstrap": "^4.3.1",
     "classnames": "^2.2.6",
     "compression": "^1.7.3",
     "connect-session-sequelize": "^6.0.0",
     "css-loader": "^2.1.1",
-    "d3": "^5.9.2",
-    "d3-scale": "^3.0.0",
+    "d3": "^5.12.0",
+    "d3-scale": "^3.2.0",
     "d3-shape": "^1.3.5",
     "daterangepicker": "^3.0.5",
-    "dotenv": "^8.0.0",
-    "express": "^4.16.4",
-    "express-session": "^1.15.1",
-    "history": "^4.9.0",
+    "dotenv": "^8.2.0",
+    "express": "^4.17.1",
+    "express-session": "^1.17.0",
+    "history": "^4.10.1",
     "lodash.debounce": "^4.0.8",
     "moment": "^2.24.0",
     "morgan": "^1.9.1",
     "passport": "^0.4.0",
     "passport-google-oauth": "^2.0.0",
-    "pg": "^7.9.0",
-    "pg-hstore": "^2.3.2",
+    "pg": "^7.12.1",
+    "pg-hstore": "^2.3.3",
     "pretty-ms": "^5.0.0",
     "prop-types": "^15.7.2",
-    "react": "^16.8.6",
-    "react-beautiful-dnd": "^11.0.2",
-    "react-csv": "^1.1.1",
-    "react-datepicker": "^2.5.0",
-    "react-dates": "^20.1.0",
-    "react-day-picker": "^7.3.0",
-    "react-dom": "^16.8.6",
-    "react-redux": "^7.0.1",
-    "react-router-dom": "^5.0.0",
-    "react-table": "^6.10.0",
-    "reactstrap": "^8.0.0",
-    "redux": "^4.0.1",
+    "react": "^16.11.0",
+    "react-beautiful-dnd": "^11.0.5",
+    "react-csv": "^1.1.2",
+    "react-datepicker": "^2.10.0",
+    "react-dates": "^20.3.0",
+    "react-day-picker": "^7.4.0",
+    "react-dom": "^16.11.0",
+    "react-redux": "^7.1.3",
+    "react-router-dom": "^5.1.2",
+    "react-table": "^6.10.3",
+    "reactstrap": "^8.1.1",
+    "redux": "^4.0.4",
     "redux-logger": "^3.0.6",
     "redux-thunk": "^2.3.0",
-    "sequelize": "^5.3.1",
-    "socket.io": "^2.2.0",
+    "sequelize": "^5.21.2",
+    "socket.io": "^2.3.0",
     "style-loader": "^0.23.1",
-    "xlsx": "^0.14.3"
+    "xlsx": "^0.14.5"
   },
   "devDependencies": {
-    "@babel/core": "^7.4.3",
+    "@babel/core": "^7.7.2",
     "@babel/plugin-proposal-class-properties": "7.4.0",
     "@babel/plugin-proposal-decorators": "7.4.0",
     "@babel/plugin-proposal-export-namespace-from": "7.2.0",
@@ -91,32 +91,32 @@
     "@babel/plugin-proposal-throw-expressions": "7.2.0",
     "@babel/plugin-syntax-dynamic-import": "7.2.0",
     "@babel/plugin-syntax-import-meta": "7.2.0",
-    "@babel/polyfill": "^7.4.3",
-    "@babel/preset-env": "^7.4.3",
-    "@babel/preset-react": "^7.0.0",
-    "@babel/register": "^7.4.0",
-    "axios-mock-adapter": "^1.16.0",
-    "babel-eslint": "^10.0.1",
-    "babel-loader": "^8.0.5",
+    "@babel/polyfill": "^7.7.0",
+    "@babel/preset-env": "^7.7.1",
+    "@babel/preset-react": "^7.7.0",
+    "@babel/register": "^7.7.0",
+    "axios-mock-adapter": "^1.17.0",
+    "babel-eslint": "^10.0.3",
+    "babel-loader": "^8.0.6",
     "chai": "^4.2.0",
-    "enzyme": "^3.9.0",
-    "enzyme-adapter-react-16": "^1.12.1",
+    "enzyme": "^3.10.0",
+    "enzyme-adapter-react-16": "^1.15.1",
     "eslint": "^5.16.0",
     "eslint-config-fullstack": "^6.0.0",
-    "eslint-config-prettier": "^4.1.0",
-    "eslint-plugin-react": "^7.12.4",
+    "eslint-config-prettier": "^4.3.0",
+    "eslint-plugin-react": "^7.16.0",
     "husky": "^1.3.1",
-    "lint-staged": "^8.1.5",
-    "mocha": "^6.1.4",
-    "nodemon": "^1.18.3",
+    "lint-staged": "^8.2.1",
+    "mocha": "^6.2.2",
+    "nodemon": "^1.19.4",
     "npm-merge-driver": "^2.3.5",
     "prettier": "1.11.1",
-    "react-test-renderer": "^16.4.2",
+    "react-test-renderer": "^16.11.0",
     "redux-devtools-extension": "^2.13.5",
     "redux-mock-store": "^1.5.3",
     "supertest": "^4.0.2",
-    "webpack": "^4.16.4",
-    "webpack-cli": "^3.1.0",
+    "webpack": "^4.41.2",
+    "webpack-cli": "^3.3.10",
     "webpack-livereload-plugin": "^2.2.0"
   }
 }
