@@ -1,5 +1,5 @@
 import React from 'react';
-import { ProjectBoard } from './project-board';
+import { ProjectBoard } from './ProjectBoard';
 import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
 import {
