@@ -1,57 +1,57 @@
-export function handleDrag(source, destination, draggableId, state) {
-  const start = state.columns[source.droppableId];
-  const finish = state.columns[destination.droppableId];
+// export function handleDrag(source, destination, draggableId, state) {
+//   const start = state.columns[source.droppableId];
+//   const finish = state.columns[destination.droppableId];
 
-  if (start === finish) {
-    const newTaskIds = Array.from(start.taskIds);
-    newTaskIds.splice(source.index, 1);
-    newTaskIds.splice(destination.index, 0, draggableId);
+//   if (start === finish) {
+//     const newTaskIds = Array.from(start.taskIds);
+//     newTaskIds.splice(source.index, 1);
+//     newTaskIds.splice(destination.index, 0, draggableId);
 
-    const newColumn = {
-      id: source.droppableId,
-      taskIds: newTaskIds
-    };
-    const newState = {
-      ...state,
-      columns: {
-        ...state.columns,
-        [newColumn.id]: newColumn
-      }
-    };
+//     const newColumn = {
+//       id: source.droppableId,
+//       taskIds: newTaskIds
+//     };
+//     const newState = {
+//       ...state,
+//       columns: {
+//         ...state.columns,
+//         [newColumn.id]: newColumn
+//       }
+//     };
 
-    return newState;
-  }
+//     return newState;
+//   }
 
-  const startTaskIds = Array.from(start.taskIds);
+//   const startTaskIds = Array.from(start.taskIds);
 
-  startTaskIds.splice(source.index, 1);
-  const newStart = {
-    id: source.droppableId,
-    taskIds: startTaskIds
-  };
+//   startTaskIds.splice(source.index, 1);
+//   const newStart = {
+//     id: source.droppableId,
+//     taskIds: startTaskIds
+//   };
 
-  const finishTaskIds = Array.from(finish.taskIds);
-  finishTaskIds.splice(destination.index, 0, draggableId);
-  const newFinish = {
-    id: destination.droppableId,
-    taskIds: finishTaskIds
-  };
+//   const finishTaskIds = Array.from(finish.taskIds);
+//   finishTaskIds.splice(destination.index, 0, draggableId);
+//   const newFinish = {
+//     id: destination.droppableId,
+//     taskIds: finishTaskIds
+//   };
 
-  const newState = {
-    ...state,
-    columns: {
-      ...state.columns,
-      [newStart.id]: newStart,
-      [newFinish.id]: newFinish
-    }
-  };
+//   const newState = {
+//     ...state,
+//     columns: {
+//       ...state.columns,
+//       [newStart.id]: newStart,
+//       [newFinish.id]: newFinish
+//     }
+//   };
 
-  return newState;
-}
+//   return newState;
+// }
 
 export function handleDragProps(source, destination, draggableId, props) {
-  const start = props.columns[source.droppableId];
-  const finish = props.columns[destination.droppableId];
+  const start = props.llColumns[source.droppableId];
+  const finish = props.llColumns[destination.droppableId];
 
   if (start === finish) {
     const newTaskIds = Array.from(start.taskIds);
@@ -65,7 +65,7 @@ export function handleDragProps(source, destination, draggableId, props) {
     const newState = {
       ...props,
       columns: {
-        ...props.columns,
+        ...props.llColumns,
         [newColumn.id]: newColumn
       }
     };
@@ -91,7 +91,7 @@ export function handleDragProps(source, destination, draggableId, props) {
   const newState = {
     ...props,
     columns: {
-      ...props.columns,
+      ...props.llColumns,
       [newStart.id]: newStart,
       [newFinish.id]: newFinish
     }
@@ -108,32 +108,45 @@ export function createTicketsObject(tickets) {
   return obj;
 }
 
-export function generateNewState(props) {
-  const newState = {
-    columns: {
-      to_do: {
-        id: 'to_do',
-        taskIds: props.to_do
-      },
-      in_progress: {
-        id: 'in_progress',
-        taskIds: props.in_progress
-      },
-      in_review: {
-        id: 'in_review',
-        taskIds: props.in_review
-      },
-      done: {
-        id: 'done',
-        taskIds: props.done
-      }
-    },
-    tickets: createTicketsObject(props.allTickets)
-  };
-  return newState;
+// export function generateNewState(props) {
+//   const newState = {
+//     columns: {
+//       to_do: {
+//         id: 'to_do',
+//         taskIds: props.to_do
+//       },
+//       in_progress: {
+//         id: 'in_progress',
+//         taskIds: props.in_progress
+//       },
+//       in_review: {
+//         id: 'in_review',
+//         taskIds: props.in_review
+//       },
+//       done: {
+//         id: 'done',
+//         taskIds: props.done
+//       }
+//     },
+//     tickets: createTicketsObject(props.allTickets)
+//   };
+//   return newState;
+// }
+
+export function generateNewColumnsLL(payload) {
+  let columns = {};
+  for (let key in payload) {
+    columns[key] = {
+      id: key,
+      name: payload[key].name,
+      taskIds: payload[key].linkedList.map(x => x.id)
+    };
+  }
+  return columns;
 }
 
 export function generateNewColumns(payload) {
+  console.log(payload);
   return {
     to_do: {
       id: 'to_do',
