@@ -1,4 +1,5 @@
 const Sequelize = require('sequelize');
+const Column = require('./column');
 const db = require('../db');
 
 const Project = db.define('project', {
@@ -6,13 +7,24 @@ const Project = db.define('project', {
     type: Sequelize.STRING,
     allowNull: false
   },
-
   totalTime: {
     type: Sequelize.INTEGER,
     validate: {
       notEmpty: true,
       min: 0
     }
+  }
+});
+
+Project.afterCreate(async project => {
+  const columns = [
+    await Column.create({ name: 'To Do' }),
+    await Column.create({ name: 'In Progress' }),
+    await Column.create({ name: 'In Review' }),
+    await Column.create({ name: 'Done' })
+  ];
+  for (let i = 0; i < columns.length; i++) {
+    await columns[i].setProject(project);
   }
 });
 
